Move collect-transform-stream into its own module.
