+++ api/get_map.js
@@ -1,8 +1,8 @@
 var collect = require('collect-stream')
 var pumpify = require('pumpify')
 var mapStream = require('through2-map')
-var through = require('through2')
 var fromArray = require('from2-array')
+var collectTransform = require('../lib/collect-transform-stream')
 
 var cmpFork = require('../lib/util').cmpFork
 
@@ -107,27 +107,4 @@ function filterForkedElements (elements) {
 var typeOrder = { node: 0, way: 1, relation: 2 }
 function cmpType (a, b) {
   return typeOrder[a.type] - typeOrder[b.type]
-}
-
-// Transform stream that performs 'fn' on an array of all elements in the
-// stream, and then passes on the resulting array onto the other end of the
-// stream.
-function collectTransform (fn) {
-  var elements = []
-
-  function write (elm, enc, next) {
-    elements.push(elm)
-    next()
-  }
-
-  function end (flush) {
-    elements = fn(elements)
-    var that = this
-    elements.forEach(function (elm) {
-      that.push(elm)
-    })
-    flush()
-  }
-
-  return through.obj(write, end)
 }
